removed deprecated Board component and event-listener.js and moved eventNames to utils
